perf(desktop): keep O(members) scans off the composer keystroke path
useCanAddChannelMembers sits in the composer send flow, which
re-renders per keystroke, and it re-scanned the channel list and the
full roster on every render. Memoize on the query data identities
(structural sharing keeps them stable), and share the roster pubkey-set
derivation with the timeline's identity-keyed cache so the Set is built
once per distinct roster rather than once per consumer.

Mention candidate construction is deliberately untouched: deferring it
changes composer highlight timing, and its memo already only recomputes
on input changes, not keystrokes.
